Added cargo feature to enable runtime module-selection
Provides a non-default cargo feature `cli-options` that enables the
ability for the user to select what modules they need at runtime. This
fits my needs where I'd like to ship the same binary to a range of hosts
but enable different modules depending on the host it's running on.
Looking at the flamegraphs (Running on ubuntu focal on docker, in a
range of contexts / git repos etc), this simple implementation doesn't
substantially impact the execution time, and shouldn't even affect the
runtime of the default build config.

There's possibly a more efficient way of optimizing the flag parsing,
but this seems good enough for the moment.